AD-220 Added refactor from Develop
