+++ src/test/java/software/amazon/documentdb/jdbc/DocumentDbStatementBasicTest.java
@@ -18,7 +18,9 @@ package software.amazon.documentdb.jdbc;
 
 import org.bson.BsonDateTime;
 import org.bson.BsonDocument;
+import org.bson.BsonTimestamp;
 import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -30,6 +32,8 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 import java.sql.Timestamp;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
@@ -526,6 +530,284 @@ public class DocumentDbStatementBasicTest extends DocumentDbStatementTest {
         Assertions.assertEquals(5, resultSet.getInt(8));
         // Seconds is 6.
         Assertions.assertEquals(6, resultSet.getInt(9));
+        Assertions.assertFalse(resultSet.next());
+    }
+
+
+    /**
+     * Tests that queries containing ORDER BY and OFFSET work.
+     * @throws SQLException occurs if query or connection fails.
+     */
+    @Test
+    @DisplayName("Tests queries with OFFSET")
+    void testQueryOffset() throws SQLException {
+        final String tableName = "testQueryOffset";
+        final BsonDocument doc1 = BsonDocument.parse("{\"_id\": 101,\n" +
+                "\"fieldA\": 1,\n" +
+                "\"fieldB\": 2}");
+        final BsonDocument doc2 = BsonDocument.parse("{\"_id\": 102,\n" +
+                "\"fieldA\": 3,\n" +
+                "\"fieldB\": 4}");
+        final BsonDocument doc3 = BsonDocument.parse("{\"_id\": 103,\n" +
+                "\"fieldA\": 5,\n" +
+                "\"fieldB\": 6}");
+        insertBsonDocuments(tableName, DATABASE_NAME, USER, PASSWORD,
+                new BsonDocument[]{doc1, doc2, doc3});
+        final Statement statement = getDocumentDbStatement();
+        final ResultSet resultSet = statement.executeQuery(
+                String.format(
+                        "SELECT * FROM \"%s\".\"%s\" LIMIT 2 OFFSET 1",
+                        DATABASE_NAME, tableName));
+        Assertions.assertNotNull(resultSet);
+        Assertions.assertTrue(resultSet.next());
+        Assertions.assertEquals("102", resultSet.getString(1));
+        Assertions.assertTrue(resultSet.next());
+        Assertions.assertEquals("103", resultSet.getString(1));
+        Assertions.assertFalse(resultSet.next());
+    }
+
+    /**
+     * Tests that queries containing IN (c1, c2...) work.
+     * @throws SQLException occurs if query or connection fails.
+     */
+    @Test
+    @DisplayName("Tests queries with WHERE [field] IN (...)")
+    void testQueryWhereIN() throws SQLException {
+        final String tableName = "testQueryWhereIN";
+        final BsonDocument doc1 = BsonDocument.parse("{\"_id\": 101,\n" +
+                "\"fieldA\": 1,\n" +
+                "\"fieldB\": \"abc\"}");
+        final BsonDocument doc2 = BsonDocument.parse("{\"_id\": 102,\n" +
+                "\"fieldA\": 3,\n" +
+                "\"fieldB\": \"def\"}");
+        final BsonDocument doc3 = BsonDocument.parse("{\"_id\": 103,\n" +
+                "\"fieldA\": 5,\n" +
+                "\"fieldB\": \"ghi\"}");
+        insertBsonDocuments(tableName, DATABASE_NAME, USER, PASSWORD,
+                new BsonDocument[]{doc1, doc2, doc3});
+        final Statement statement = getDocumentDbStatement();
+        ResultSet resultSet = statement.executeQuery(
+                String.format(
+                        "SELECT * FROM \"%s\".\"%s\" WHERE \"fieldA\" IN (1, 5)",
+                        DATABASE_NAME, tableName));
+        Assertions.assertNotNull(resultSet);
+        Assertions.assertTrue(resultSet.next());
+        Assertions.assertEquals("101", resultSet.getString(1));
+        Assertions.assertTrue(resultSet.next());
+        Assertions.assertEquals("103", resultSet.getString(1));
+        Assertions.assertFalse(resultSet.next());
+        resultSet = statement.executeQuery(
+                String.format(
+                        "SELECT * FROM \"%s\".\"%s\" WHERE \"fieldB\" IN ('abc', 'ghi')",
+                        DATABASE_NAME, tableName));
+        Assertions.assertNotNull(resultSet);
+        Assertions.assertTrue(resultSet.next());
+        Assertions.assertEquals("101", resultSet.getString(1));
+        Assertions.assertTrue(resultSet.next());
+        Assertions.assertEquals("103", resultSet.getString(1));
+        Assertions.assertFalse(resultSet.next());
+    }
+
+    /**
+     * Tests that queries containing NOT IN (c1, c2...) work.
+     * @throws SQLException occurs if query or connection fails.
+     */
+    @Test
+    @DisplayName("Tests queries with WHERE [field] NOT IN (...)")
+    void testQueryWhereNotIN() throws SQLException {
+        final String tableName = "testQueryWhereNOTIN";
+        final BsonDocument doc1 = BsonDocument.parse("{\"_id\": 101,\n" +
+                "\"fieldA\": 1,\n" +
+                "\"fieldB\": \"abc\"}");
+        final BsonDocument doc2 = BsonDocument.parse("{\"_id\": 102,\n" +
+                "\"fieldA\": 3,\n" +
+                "\"fieldB\": \"def\"}");
+        final BsonDocument doc3 = BsonDocument.parse("{\"_id\": 103,\n" +
+                "\"fieldA\": 5, \n" +
+                "\"fieldB\": \"ghi\"}");
+        insertBsonDocuments(tableName, DATABASE_NAME, USER, PASSWORD,
+                new BsonDocument[]{doc1, doc2, doc3});
+        final Statement statement = getDocumentDbStatement();
+        ResultSet resultSet = statement.executeQuery(
+                String.format(
+                        "SELECT * FROM \"%s\".\"%s\" WHERE \"fieldA\" NOT IN (1, 5)",
+                        DATABASE_NAME, tableName));
+        Assertions.assertNotNull(resultSet);
+        Assertions.assertTrue(resultSet.next());
+        Assertions.assertEquals("102", resultSet.getString(1));
+        Assertions.assertFalse(resultSet.next());
+        resultSet = statement.executeQuery(
+                String.format(
+                        "SELECT * FROM \"%s\".\"%s\" WHERE \"fieldB\" NOT IN ('abc', 'ghi')",
+                        DATABASE_NAME, tableName));
+        Assertions.assertNotNull(resultSet);
+        Assertions.assertTrue(resultSet.next());
+        Assertions.assertEquals("102", resultSet.getString(1));
+        Assertions.assertFalse(resultSet.next());
+    }
+
+    /**
+     * Tests that queries containing casts from numbers work.
+     * @throws SQLException occurs if query or connection fails.
+     */
+    @Test
+    @DisplayName("Tests queries with cast from number.")
+    void testQueryCastNum() throws SQLException {
+        final String tableName = "testQueryCastNum";
+        final BsonDocument doc1 = BsonDocument.parse("{\"_id\": 101,\n" +
+                "\"fieldA\": 1,\n" +
+                "\"fieldB\": 1.2," +
+                "\"fieldC\": 10000000000}");
+        insertBsonDocuments(tableName, DATABASE_NAME, USER, PASSWORD,
+                new BsonDocument[]{doc1});
+        final Statement statement = getDocumentDbStatement();
+        final ResultSet resultSet = statement.executeQuery(
+                String.format(
+                        "SELECT CAST(\"fieldA\" AS INTEGER), " +
+                                "CAST(\"fieldA\" AS BIGINT), " +
+                                "CAST(\"fieldA\" AS DOUBLE), " +
+                                "CAST(\"fieldB\" AS INTEGER)," +
+                                "CAST(\"fieldB\" AS BIGINT), " +
+                                "CAST(\"fieldB\" AS DOUBLE), " +
+                                "CAST(\"fieldC\" AS BIGINT), " +
+                                "CAST(\"fieldC\" AS DOUBLE) " +
+                                " FROM \"%s\".\"%s\"",
+                        DATABASE_NAME, tableName));
+        Assertions.assertNotNull(resultSet);
+        Assertions.assertTrue(resultSet.next());
+        Assertions.assertEquals(1, resultSet.getInt(1));
+        Assertions.assertEquals(1L, resultSet.getLong(2));
+        Assertions.assertEquals(1D, resultSet.getDouble(3));
+        Assertions.assertEquals(1, resultSet.getInt(4));
+        Assertions.assertEquals(1L, resultSet.getLong(5));
+        Assertions.assertEquals(1.2D, resultSet.getDouble(6));
+        Assertions.assertEquals(10000000000L, resultSet.getLong(7));
+        Assertions.assertEquals(10000000000D, resultSet.getDouble(8));
+        Assertions.assertFalse(resultSet.next());
+    }
+
+    /**
+     * Tests that queries containing casts from strings work.
+     * @throws SQLException occurs if query or connection fails.
+     */
+    @Test
+    @Disabled("Cast to date not working.")
+    @DisplayName("Tests queries with cast from string.")
+    void testQueryCastString() throws SQLException, ParseException {
+        final String tableName = "testQueryCastString";
+        final BsonDocument doc1 = BsonDocument.parse("{\"_id\": 101,\n" +
+                "\"fieldA\": \"2020-03-11\", \n" +
+                "\"fieldNum\": \"100.5\"}");
+        insertBsonDocuments(tableName, DATABASE_NAME, USER, PASSWORD,
+                new BsonDocument[]{doc1});
+        final Statement statement = getDocumentDbStatement();
+        final ResultSet resultSet = statement.executeQuery(
+                String.format(
+                        "SELECT CAST(\"fieldA\" AS DATE), " +
+                                "CAST(\"fieldA\" AS TIMESTAMP), " +
+                                "CAST(\"fieldNum\" AS DOUBLE), " +
+                                "CAST(\"fieldNum\" AS INTEGER)," +
+                                "CAST(\"fieldNum\" AS BIGINT) " +
+                                " FROM \"%s\".\"%s\"",
+                        DATABASE_NAME, tableName));
+        Assertions.assertNotNull(resultSet);
+        Assertions.assertTrue(resultSet.next());
+        Assertions.assertEquals("2020-03-11", resultSet.getString(1));
+        Assertions.assertEquals(new SimpleDateFormat("yyyy/MM/dd").parse("2020/03/11").getTime(),
+                resultSet.getTimestamp(2).getTime());
+        Assertions.assertEquals(100.5D, resultSet.getDouble(3));
+        Assertions.assertEquals(100, resultSet.getInt(4));
+        Assertions.assertEquals(100, resultSet.getLong(5));
+        Assertions.assertFalse(resultSet.next());
+    }
+
+    /**
+     * Tests that queries containing casts from dates work.
+     * @throws SQLException occurs if query or connection fails.
+     */
+    @Test
+    @Disabled("Cast to date not working.")
+    @DisplayName("Tests queries with cast from date.")
+    void testQueryCastDate() throws SQLException, ParseException {
+        final String tableName = "testQueryCastDate";
+        final BsonDocument doc1 = BsonDocument.parse("{\"_id\": 101}");
+        doc1.append("date",
+                new BsonTimestamp(new SimpleDateFormat("yyyy/MM/dd").parse("2020/03/11").getTime()));
+        insertBsonDocuments(tableName, DATABASE_NAME, USER, PASSWORD,
+                new BsonDocument[]{doc1});
+        final Statement statement = getDocumentDbStatement();
+        final ResultSet resultSet = statement.executeQuery(
+                String.format(
+                        "SELECT CAST(\"date\" AS DATE), " +
+                                "CAST(\"date\" AS TIMESTAMP), " +
+                                "CAST(\"date\" AS TIME)," +
+                                "CAST(\"date\" AS VARCHAR) " +
+                                " FROM \"%s\".\"%s\"",
+                        DATABASE_NAME, tableName));
+        Assertions.assertNotNull(resultSet);
+        Assertions.assertTrue(resultSet.next());
+        Assertions.assertEquals("2020-03-11", resultSet.getDate(1).toString());
+        Assertions.assertEquals(new SimpleDateFormat("yyyy/MM/dd").parse("2020/03/11").getTime(),
+                resultSet.getTimestamp(2).getTime());
+        Assertions.assertEquals(new SimpleDateFormat("yyyy/MM/dd").parse("2020/03/11").getTime(),
+                resultSet.getTime(3).getTime());
+        Assertions.assertFalse(resultSet.next());
+    }
+
+    /**
+     * Tests that queries containing casts from boolean work.
+     * @throws SQLException occurs if query or connection fails.
+     */
+    @Test
+    @DisplayName("Tests queries with cast from boolean.")
+    void testQueryCastBoolean() throws SQLException {
+        final String tableName = "testQueryCastBoolean";
+        final BsonDocument doc1 = BsonDocument.parse("{\"_id\": 101,\n" +
+                "\"fieldA\": false}");
+        insertBsonDocuments(tableName, DATABASE_NAME, USER, PASSWORD,
+                new BsonDocument[]{doc1});
+        final Statement statement = getDocumentDbStatement();
+        final ResultSet resultSet = statement.executeQuery(
+                String.format(
+                        "SELECT CAST(\"fieldA\" AS VARCHAR)" +
+                                " FROM \"%s\".\"%s\"",
+                        DATABASE_NAME, tableName));
+        Assertions.assertNotNull(resultSet);
+        Assertions.assertTrue(resultSet.next());
+        Assertions.assertEquals("false", resultSet.getString(1));
+        Assertions.assertFalse(resultSet.next());
+    }
+
+    /**
+     * Tests that queries containing various nested casts work.
+     * @throws SQLException occurs if query or connection fails.
+     */
+    @Test
+    @Disabled("Casts are not functioning from string to date.")
+    @DisplayName("Tests queries with nested CAST.")
+    void testQueryNestedCast() throws SQLException, ParseException {
+        final String tableName = "testQueryNestedCast";
+        final BsonDocument doc1 = BsonDocument.parse("{\"_id\": 101,\n" +
+                "\"dateString\": \"2020-03-11\"," +
+                "\"fieldNum\": 7," +
+                "\"fieldString\": \"5\"}");
+        doc1.append("fieldTimestamp", new BsonTimestamp(
+                new SimpleDateFormat("yyyy/MM/dd").parse("2020/03/11").getTime()));
+        insertBsonDocuments(tableName, DATABASE_NAME, USER, PASSWORD,
+                new BsonDocument[]{doc1});
+        final Statement statement = getDocumentDbStatement();
+        final ResultSet resultSet = statement.executeQuery(
+                String.format(
+                        "SELECT CAST(CAST(\"dateString\" AS DATE) AS VARCHAR), " +
+                                "CAST(CAST(\"fieldTimestamp\" AS DATE) AS VARCHAR), " +
+                                "CAST(CAST(\"fieldNum\" AS DOUBLE) AS INTEGER)" +
+                                " FROM \"%s\".\"%s\"",
+                        DATABASE_NAME, tableName));
+        Assertions.assertNotNull(resultSet);
+        Assertions.assertTrue(resultSet.next());
+        Assertions.assertEquals("2020-03-11", resultSet.getString(1));
+        Assertions.assertEquals("2020-03-11", resultSet.getString(2));
+        Assertions.assertEquals(5, resultSet.getInt(3));
         Assertions.assertFalse(resultSet.next());
     }
 }
